+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 
 - Prevent non-compliant hostnames by replacing underscores with dashes
+- use shadowenv for configuring terraform backend secrets
+- use GitLab http backend for terraform state storage whenever configuring both terraform and GitLab CI
 
 ## [2.1.0] - 2025-04-29